+++ src/tables/mod.rs
@@ -11,3 +11,4 @@ pub mod hhea;
 pub mod hmtx;
 pub mod maxp;
 pub mod name;
+pub mod post;